Reorganise components
This rearranges the components into a more standard interface that can be applied to all UI primitives:
{EditComponent, ViewComponent, parser}

Also splits out Relationship.jsx into a separate component file so it is more testable when we come to enhance it with properties on relationships
