+++ packages/tc-ui/relationship/browser.jsx
@@ -2,7 +2,7 @@ const { render, h } = require('preact');
 const { RelationshipPicker } = require('./lib/relationship-picker.jsx');
 
 module.exports = {
-	attachEditComponent: container =>
+	withEditComponent: container =>
 		render(
 			<RelationshipPicker {...JSON.parse(container.dataset.props)} />,
 			document,
